Updates composer file
Updates composer file to add the supported range of PHP versions, and updates, removes or adds the dependencies of the SDK.

Supported PHP versions start from 5.6.4.
Removes the guzzle dependency, making it more flexible when it comes to use the SDK in different PHP versions.
Defines two versions of phpunit for unit testing, in a way that composer chooses the version that fits better to the PHP version in use.